feat: fully migrated operators to AST
