+++ compiler/ast_parser/src/types.rs
@@ -47,7 +47,7 @@ pub fn parse_type_type_parameters(tokens: &Vec<LexerToken>, ind: &mut usize) -> 
 
 		types.push(parsed_type);
 
-		if tokens[*ind].is_angel_bracket_close() {
+		if tokens[*ind].tok_type == LexerTokenType::AngelBracketClose {
 			break;
 		}
 
@@ -206,7 +206,7 @@ pub fn parse_type_parameters_declaration(tokens: &Vec<LexerToken>, ind: &mut usi
 
 		*ind += 1;
 		
-		if tokens[*ind].is_angel_bracket_close() {
+		if tokens[*ind].tok_type == LexerTokenType::AngelBracketClose {
 			break;
 		}
