+++ compiler/astoir_mir_lowering/Cargo.toml
@@ -7,6 +7,5 @@ edition = "2024"
 astoir_mir = {path = "../astoir_mir"}
 astoir_hir = {path = "../astoir_hir"}
 diagnostics = { path = "../diagnostics" }
-lexer = {path = "../lexer"}
 compiler_typing = { path = "../compiler_typing" }
 compiler_utils = { path = "../compiler_utils" }
